Refactor ResolvePossibleSymlinkToHostPath to common directory
